requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
ecdsa==0.18.0
exceptiongroup==1.0.4
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
packaging==22.0
phonenumbers==8.13.3
pluggy==1.0.0
psycopg2==2.9.5
pyasn1==0.4.8
pycountry==22.3.5
pycparser==2.21
pydantic==1.10.2
pytest==7.2.0
pytest-asyncio==0.20.3
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
